externalizes utils to `./lib/utils.js`, adds prettify helper to make diffs easier to review.
