+++ README.md
@@ -546,7 +546,7 @@ Released under the MIT license
 
 ***
 
-_This file was generated on Wednesday, October 16, 2013._
+_This file was generated on Thursday, October 17, 2013._
 
 
 [moment]: http://momentjs.com/ "Moment.js Permalinks"
