+++ package.json
@@ -33,7 +33,8 @@
     "assemble": "~0.4.12",
     "grunt-contrib-clean": "~0.5.0",
     "grunt-contrib-jshint": "~0.6.4",
-    "grunt-readme": "~0.1.7"
+    "grunt-readme": "~0.1.7",
+    "helper-prettify": "~0.1.7"
   },
   "peerDependencies": {
     "assemble": "~0.4.12"
